Style ipython notebooks.
This grows our minimized blog.css to almost 200k which is not
great. We'd probably be better off compiling a smaller, custom CSS
from IPython's less files but for now this will do.

I'm also employing a pretty bad hack to make sure that the IPython css
only styles the parts of the page we want: I'm including their
minimized css in a subelement of our scss.

There are a few ugly things, e.g. IPython's css sets a line-height of
20px which I can't override easily.